requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
hashids>=1.2.0,<1.3.0
django-cleanup>=4.0.0,<4.1.0
Pillow>=7.0.0,<7.1.0